RAG/requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain_chroma
chromadb
pypdf
sentence-transformers
python-dotenv